refactor: BridgeGame에서 수행하던 기능을 BridgeMap으로 이동
- BridgeMap에서 지도 업데이트 기능을 수행하도록 변경함
